Moved *Requests tests to extrequests
