+++ amy/extrequests/tests/test_event_submissions.py
@@ -1,11 +1,8 @@
-import unittest
-
-from django.core import mail
 from django.urls import reverse
 
-from workshops.tests.base import TestBase
-from workshops.forms import EventSubmitForm
+from extforms.deprecated_forms import EventSubmitForm
 from workshops.models import EventSubmission, Organization, Tag, Event
+from workshops.tests.base import TestBase
 
 
 class TestEventSubmitForm(TestBase):
@@ -25,50 +22,6 @@ class TestEventSubmitForm(TestBase):
             'self_organized', 'notes', 'privacy_consent', 'captcha',
         ]
         self.assertEqual(fields_left, fields_right)
-
-    @unittest.skip('Event submission disabled as per @maneesha\'s request.')
-    def test_submission_added(self):
-        """Test if the submitted form adds a new event submission."""
-        self.assertEqual(len(EventSubmission.objects.all()), 1)
-        data = {
-            'contact_name': 'Harry Potter',
-            'contact_email': 'harry@potter.com',
-            'self_organized': True,
-            'url': 'http://example.org/2016-02-13-Howart',
-            'notes': '',
-            'privacy_consent': True,
-        }
-        self.passCaptcha(data)
-
-        rv = self.client.post(reverse('event_submit'), data,
-                              follow=True)
-        self.assertEqual(rv.status_code, 200)
-        self.assertNotIn('form', rv.context)
-        self.assertEqual(len(EventSubmission.objects.all()), 2)
-
-    @unittest.skip('Event submission disabled as per @maneesha\'s request.')
-    def test_submission_sends_email(self):
-        """Test if the submitted form results in email sent."""
-        data = {
-            'contact_name': 'Harry Potter',
-            'contact_email': 'harry@potter.com',
-            'self_organized': True,
-            'url': 'http://example.org/2016-02-13-Hogwart',
-            'notes': '',
-            'privacy_consent': True,
-        }
-        self.passCaptcha(data)
-
-        rv = self.client.post(reverse('event_submit'), data,
-                              follow=True)
-        self.assertEqual(rv.status_code, 200)
-        self.assertNotIn('form', rv.context)
-        self.assertEqual(len(mail.outbox), 1)
-        msg = mail.outbox[0]
-        self.assertEqual(
-            msg.subject,
-            'New workshop submission from Harry Potter'
-        )
 
     def test_submission_accepted_with_event(self):
         """Ensure submission is turned inactive after acceptance."""
@@ -112,7 +65,7 @@ class TestEventSubmitForm(TestBase):
     def test_discarded_submission_reopened(self):
         self.submission.state = "d"
         self.submission.save()
-        rv = self.client.get(
+        self.client.get(
             reverse('eventsubmission_set_state',
                     args=[self.submission.pk, 'pending']),
             follow=True)
@@ -122,7 +75,7 @@ class TestEventSubmitForm(TestBase):
     def test_accepted_submission_reopened(self):
         self.submission.state = "a"
         self.submission.save()
-        rv = self.client.get(
+        self.client.get(
             reverse('eventsubmission_set_state',
                     args=[self.submission.pk, 'pending']),
             follow=True)
