+++ amy/extrequests/tests/test_training_request.py
